+++ packages/jest-transform/src/types.ts
@@ -22,6 +22,8 @@ export type Options = ShouldInstrumentOptions &
   Partial<{
     isCoreModule: boolean;
     isInternalModule: boolean;
+    supportsDynamicImport: boolean;
+    supportsStaticESM: boolean;
   }>;
 
 // extends directly after https://github.com/sandersn/downlevel-dts/issues/33 is fixed
@@ -38,15 +40,18 @@ export type TransformedSource =
 
 export type TransformResult = TransformTypes.TransformResult;
 
-export type TransformOptions = {
+export interface TransformOptions {
   instrument: boolean;
-};
+  // names are copied from babel
+  supportsDynamicImport?: boolean;
+  supportsStaticESM?: boolean;
+}
 
-export type CacheKeyOptions = {
+// TODO: For Jest 26 we should combine these into one options shape
+export interface CacheKeyOptions extends TransformOptions {
   config: Config.ProjectConfig;
-  instrument: boolean;
   rootDir: string;
-};
+}
 
 export interface Transformer {
   canInstrument?: boolean;